implemented Dr Jack Whitham subtle fixes for subtle rendering issues: http://blog.jwhitham.org/2015/04/gcc-bug-323-journey-to-heart-of.html

+++ ref_soft/r_edge.c
@@ -1075,7 +1075,7 @@ void D_DrawflatSurfaces (void)
 
 		// make a stable color for each surface by taking the low
 		// bits of the msurface pointer
-		D_FlatFillSurface (s, (int)s->msurf & 0xFF);
+		D_FlatFillSurface (s, (int)((intptr_t) s->msurf) & 0xFF);
 		D_DrawZSpans (s->spans);
 	}
 }
